load_testing/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
locust = "^2.27.0"
websocket-client = "^1.8.0"
rel = "^0.4.9.19"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"
mypy = "^1.10.0"
types-requests = "^2.32.0.20240521"
types-gevent = "^24.2.0.20240425"

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
